Refactor EventSub component using Router-Handler pattern
Breaks down the monolithic eventsub.js (891 lines) into specialized handlers
for better maintainability, testability, and scalability.

Architecture changes:
- eventsub.js (263 lines): Main router handling verification, deduplication,
  and event routing
- eventUtils.js (64 lines): Shared utility functions for broadcaster ID lookup
  and shared session info
- handlers/chatHandler.js (126 lines): Chat message processing logic
- handlers/redemptionHandler.js (320 lines): Channel Points redemption logic
  with validation and rejection
- handlers/notificationHandler.js (112 lines): Subs, raids, follows, cheers
- handlers/sharedChatHandler.js (81 lines): Shared chat session management

Benefits:
- Main router file reduced by 70% (891 → 263 lines)
- Single Responsibility Principle: Each handler focuses on one event category
- Easier to test individual handlers in isolation
- Simpler to add new event types without modifying core router
- Better code organization and readability

No functional changes - all business logic preserved.
